Conflicts:
	Multijoueur/src/network2/Client.java
	Multijoueur/src/network2/Connexion.java
	Multijoueur/src/network2/Server.java

+++ RPG Project/src/application/fonction/Parametres.java
@@ -2,6 +2,6 @@ package application.fonction;
 
 public class Parametres {
 
-	public static float VITESSE_DEPLACEMENT = 10.0F;
+	public static int VITESSE_DEPLACEMENT = 10;
 	
 }
